address review: drop test license header, move KEDA helper to keda.go
Remove the Apache license header from hpa_index_test.go per project
convention, and move IsOwnedByKEDAScaledObject out of utils.go into a
dedicated internal/utils/keda.go.
